requirements.txt
rootpath >= 0.1.0
attributedict >= 0.2.0
deepmerge >= 0.0.4
pyyaml >= 3.12
six >= 1.11.0

colour-runner >= 0.0.5
deepdiff >= 3.3.0
pygments >= 2.2.0
tox >= 3.0.0
